requirements.txt
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
Flask==2.2.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
outcome==1.2.0
requests==2.28.2
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
urllib3==1.26.14
wsproto==1.2.0
pip==22.3.1
setuptools==65.6.3
python-dotenv==0.21.1